Added NoLanguage, code cleanup
